Tweak to projects.py to skip session XML retrieve if the file already exists.

+++ src/xnat_cli_scripts/projects.py
@@ -2167,12 +2167,15 @@ def execute_get_session_xml(connection: XNATSession, args: argparse.Namespace) -
 
         try:
             Path(args.output_folder, project_id).mkdir(parents=True, exist_ok=True)
+            output_file = Path(args.output_folder, project_id, f"{session_id}.xml")
+            if output_file.exists():
+                print(f"{session_index} / {session_count}  Project {project_id}  Subject {subject_id}  Session {session_id}  Already exists on disk ")
+                continue
+
             print(f"{session_index} / {session_count}  Project {project_id}  Subject {subject_id}  Session {session_id}  ")
             response = connection.get(f"/data/experiments/{session_id}?format=xml")
             response.raise_for_status()
 
-#            output_file = Path(args.output_folder, project_id) / f"{session_id}.xml"
-            output_file = Path(args.output_folder, project_id, f"{session_id}.xml")
             with open(output_file, "w", encoding="utf-8") as f:
                 f.write(response.content.decode("utf-8"))
                 f.close()
